refactor: popup에서 modal로 네이밍 변경
